### 2018-11-17 9916b5d
Add "Department of Statistics" to affiliation in paper.md
