### 2018-11-17 #1086 @trinker
>  @trinker - would you be willing to review this submission for JOSS?

@arfon Yes
