### 2018-11-18 #1086 @yochannah
Hey @garretrc, thanks for submitting this paper for review! I'll be editing this for JOSS.

@trinker thanks for being willing to review!

The next step is to find a second reviewer before we kick off the official review. @trinker @garretrc - any suggestions for reviewer number 2 by any chance, perhaps from the [reviewer spreadsheet](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit?usp=sharing)? No worries if not!
