### 2018-11-18 #1086 @garretrc
Maybe @nuest or @jhollist ? Based on folks who listed R and spatial in the spreadsheet
