### 2018-11-18 #1086 @jhollist
Not a firm no.  I have some possible work changes coming up that would limit my ability to do a review.  I should know more in 10-12 days.  If you can get someone else that'd be best, but I'll keep y'all updated on my status.
