### 2018-11-19 #1086 @garretrc
Looking at people who listed R and data viz I see @corybrunson and @tweed1e if either of you would be willing to review!
