### 2018-11-19 #1086 @corybrunson
Probably—i hope to finish a review for a paper journal today or tomorrow; if i succeed, then i'll be glad to take this on and finish within a week. Will update within a day.
