### 2018-11-25 #1096 @whedon
[REVIEW]: ggvoronoi: Voronoi Diagrams and Heatmaps with ggplot2
**Submitting author:** @garretrc (<a href="http://orcid.org/0000-0002-1434-6630">Robert Garrett</a>)
**Repository:** <a href="https://github.com/garretrc/ggvoronoi" target ="_blank">https://github.com/garretrc/ggvoronoi</a>
**Version:** v0.8.2
