### 2018-11-26 #1096 @corybrunson
Thanks @garretrc for this elegant Voronoi extension! I'll start raising issues at the code repo momentarily. As you address the issues, i'll check off the related items on my review checklist. If you think some of my issues aren't really problematic, or if you think my suggestions are not great, then please let me know in the issue thread and handle them as you see fit—i'll probably then better understand the issue and check it off for reviewing purposes.
