### 2018-11-26 garretrc/ggvoronoi#1 @corybrunson
author contributions?
One author contributed code through Git, and another co-authored the vignette, though i don't doubt that all four credited authors contributed significantly to the package. To address one question from the review, could the documentation or the software paper include a brief note on each author's contribution?

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/1096)._
