### 2018-11-26 garretrc/ggvoronoi#2 @corybrunson
statement of need
The vignette provides several examples, including two real-world case studies. However, neither the `DESCRIPTION` nor the vignette include a few general sentences on the kinds of problems the package is intended to help users solve (as the software paper does).

Such purpose statement might be included in a central documentation page accessible via `?ggvoronoi`. (See `?ggplot2` for an example.)

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/1096)._
