### 2018-11-26 garretrc/ggvoronoi#3 @corybrunson
add hyperlinks in documentation
The documentation is not hyperlinked. For example:

* there are references in `?geom_voronoi` to `stat_voronoi`, but one must navigate to the index in order to find them
* other packages are not linked when referenced, e.g. when `?voronoi_polygon` refers to `geom_polygon`
* the `ncdc_locations` data set documentation includes a URL but it is not clickable

Could this be remedied?

I don't know how much this matters in general; but, because this package is targeted to students, i think navigational convenience is especially important. (I recommend [Wickham's book](http://r-pkgs.had.co.nz/man.html) for guidance, but there are other resources.)

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/1096)._
