### 2018-11-26 garretrc/ggvoronoi#4 @corybrunson
(un-)export ggproto objects
The `'ggproto'` object `StatVoronoi` is exported but not documented; should it be un-exported (like `GeomVoronoi`)? or should both be exported (and documented)?

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/1096)._
