### 2018-11-26 garretrc/ggvoronoi#5 @corybrunson
Oxford bikes data set documentation
The `oxford_bikes` data set cites four authors as a source but provides no additional information. Where did this data set come from, and how was it generated?

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/1096)._
