### 2018-11-26 garretrc/ggvoronoi#6 @corybrunson
omitted DOI for ggplot2
JOSS prefers that references include DOIs if possible. Most of the references without DOIs are R packages or blog posts (and don't have DOIs), but Wickham's book on `ggplot2` is cited as a book and therefore could helpfully include the DOI of [the Springer–Verlag edition](https://www.springer.com/us/book/9780387981413).

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/1096)._
