### 2018-11-26 5eb5e1d
added doi for ggplot2 to paper
