+++ paper.bib
@@ -58,6 +58,7 @@
   publisher = {Springer-Verlag New York},
   year = {2016},
   isbn = {978-3-319-24277-4},
+  doi = {10.1007/978-0-387-98141-3},
   url = {http://ggplot2.org},
 }
 
