### 2018-11-26 #1096 @garretrc
Thanks for the feedback! I've started to go through the issues on the ggvoronoi repository. I'll close issues on that repo if I'm confident the issue is resolved and comment if I have any questions.
