### 2018-11-26 garretrc/ggvoronoi#4 @garretrc
StatVoronoi should be exported so it can be called via stat="voronoi" in any ggplot2 geom, but GeomVoronoi is more of a convenience function: people are used to calling geom_* in ggplot rather than stat_*. We debated removing geom_voronoi entirely, as it technically works the same as the default behavior of stat_voronoi, but ultimately decided to include the function as a convenience to users.

As of now, StatVoronoi does not have its own .Rd file but calling ?StatVoronoi brings up the documentation page for stat_voronoi. ggplot2 handles calls to Stat* by bringing up the documentation page for ggplot2-ggproto, so I think the behavior of ?StatVoronoi is appropriate, unless you think an additional .Rd file is called for in this instance or have a different idea!
