### 2018-11-26 364225f
Update CONTRIBUTING.md
includes contributions of each author/contributor
