+++ CONTRIBUTING.md
@@ -8,3 +8,25 @@ Have a good idea for ggvoronoi? Let's work together! First, please check out the
 * If you have found a bug, please include information about your R version and system OS. 
 * For each pull request you want to make, please create a new branch. 
 * Please include tests for each edit or new contribution made. Testing in 'ggvoronoi' is done using 'testthat' and 'vdiffr'.
+
+## Contributions
+
+Robert Garrett: Package maintainer, co-creator, and co-author
+* Author of fortify_voronoi
+* Author of unit tests
+* Co-author for base functionality of voronoi_polygon
+* Co-author for ggplot2 integration
+* Co-author of vignette
+* Co-author of documentation
+
+Austin Nar: Co-creator and co-author
+* Co-author for base functionality of voronoi_polygon
+* Co-author for ggplot2 integration
+
+Thomas Fisher: Co-author
+* Co-author of vignette
+* Co-author of documentation
+
+Karsten Maurer: Contributor
+* Contributions to package ideation
+* Contributions to ggplot2 integration
