### 2018-11-26 44b02be
Update CONTRIBUTING.md
