+++ README.md
@@ -43,7 +43,7 @@ Alternatively, the vignette is available [here.](http://htmlpreview.github.io/?h
 
 For questions about usage/syntax or for reporting bugs, email Robert Garrett at garretrc@miamioh.edu or raise an issue on this repository. 
 
-To contribute to the package, follow the procedures outlined in CONTRIBUTING.md
+To contribute to the package or see past contributions, follow the procedures outlined in CONTRIBUTING.md
 
 ## Changes
 
