### 2018-11-26 garretrc/ggvoronoi#1 @garretrc
Added some high-level descriptions of author/contributor contributions in CONTRIBUTING.md, and included a few words in the README pointing to this change
