+++ paper.md
@@ -33,7 +33,7 @@ affiliations:
 
 Spatial data are often gathered at the individual point level, structured as a set of coordinates coupled with a set of measurements. However, attempts to visualize point-level data usually fall into two categories: a scatter plot or some spatial binning technique such as a raster or choropleth diagram of data aggregates. The Voronoi diagram, a mathematical tool that partitions a plane into localized regions around observed points, can be used to create an alternative method for point-level data visualization. A Voronoi diagram is a nearest neighbor diagram: given a set of generator points, the Voronoi diagram creates nearest neighbor cells. Each cell is formed around one generator point, and the set of points closest to that generator is fully contained in its cell. The resulting diagram can be used to find the nearest generator for an arbitrary point or to interpolate the blank space around each generator to create a heatmap.
 
-Voronoi diagrams have a multitude of applications, but ``ggvoronoi`` implements toold for data visualization and spatial data analysis in ``R``. For data visualization, ``ggvoronoi`` fits nicely within the ``ggplot2`` framework, so ``R`` users experienced with the popular visualization package can create Voronoi diagrams and heatmaps around any point-level dataset within minutes. For spatial data analysis, ``R`` users can create a Voronoi diagram within the ``sp`` framework as a ``SpatialPolygonsDataFrame`` object. From there, the user can find the nearest generator point to each location in a new ``SpatialPoints`` object or use tools, such as those contained in``rgeos``, to analyze the diagram.
+Voronoi diagrams have a multitude of applications, but ``ggvoronoi`` implements toold for data visualization and spatial data analysis in ``R``. For data visualization, ``ggvoronoi`` fits nicely within the ``ggplot2`` framework, so ``R`` users experienced with the popular visualization package can create Voronoi diagrams and heatmaps around any point-level dataset within minutes. For spatial data analysis, ``R`` users can create a Voronoi diagram within the ``sp`` framework as a ``SpatialPolygonsDataFrame`` object. From there, the user can find the nearest generator point to each location in a new ``SpatialPoints`` object or use tools, such as those contained in ``rgeos``, to analyze the diagram.
 
 ![Comparison of a colored scatter plot and Voronoi diagram heatmap. The data containing the National Climatic Data Center weather station coordinates and elevations are included in the package, and the code for this plot is in the package vignette.](california.jpeg)
 
