### 2018-11-26 80d4cb4
typo in paper.md
