### 2018-11-26 e100b24
update README.md to contain information on package purpose
