### 2018-11-26 a5defb1
added documentation for ?ggvoronoi
