### 2018-11-26 garretrc/ggvoronoi#2 @garretrc
Added some package purpose information from the paper into the readme, and created a new documentation file for ?ggvoronoi with that same information!
