### 2018-11-26 f55f31c
added links to fortify_voronoi docs:
