+++ R/fortify_voronoi.R
@@ -1,7 +1,8 @@
 #' Create Data Frame from voronoi_polygons output
-#'
-#' Allows output of voronoi_polygon to be plotted using geom_polygon.
-#' @param vor_spdf Output from voronoi_polygon (SpatialPolygonsDataFrame)
+#' 
+#' Allows output of \code{\link[ggvoronoi]{voronoi_polygon}} to be plotted using
+#' \code{\link[ggplot2]{geom_polygon}}.
+#' @param vor_spdf Output from \code{\link[ggvoronoi]{voronoi_polygon}} (\code{\link[sp]{SpatialPolygonsDataFrame}})
 #' @keywords voronoi, choropleth
 #' @import ggplot2
 #' @export
@@ -14,10 +15,10 @@
 #' circle <- data.frame(x = 100*(1+cos(seq(0, 2*pi, length.out = 2500))),
 #'                      y = 100*(1+sin(seq(0, 2*pi, length.out = 2500))),
 #'                      group = rep(1,2500))
-#'
+#' 
 #' vor_spdf <- voronoi_polygon(data=points,x="x",y="y",outline=circle)
 #' vor_df <- fortify_voronoi(vor_spdf)
-#'
+#' 
 #' ggplot(vor_df) +
 #'     geom_polygon(aes(x=x,y=y,fill=distance,group=group))
 
