### 2018-11-26 df21e03
added hyperlinks to documentation pages
