+++ man/voronoi_polygon.Rd
@@ -8,18 +8,18 @@ voronoi_polygon(data, x = "x", y = "y", outline = NULL,
   data.frame = FALSE)
 }
 \arguments{
-\item{data}{data.frame containing a set of points to make voronoi regions and any additional desired columns.}
+\item{data}{\code{data.frame} containing a set of points to make voronoi regions and any additional desired columns.}
 
 \item{x}{numeric vector (for example longitude).}
 
 \item{y}{numeric vector (for example latitude).}
 
-\item{outline}{data.frame with first column x/longitude, second column y/latitude, and a group column denoting islands or pieces.}
+\item{outline}{\code{data.frame} with first column x/longitude, second column y/latitude, and a group column denoting islands or pieces.}
 
-\item{data.frame}{output as data.frame? You will lose information if you do this. For use in StatVoronoi.}
+\item{data.frame}{output as \code{data.frame}? You will lose information if you do this. For use in \code{\link[ggvoronoi]{StatVoronoi}}.}
 }
 \description{
-Create a Voronoi diagram for analysis or plotting with geom_polygon.
+Create a Voronoi diagram for analysis or plotting with \code{\link[ggplot2]{geom_polygon}}.
 }
 \examples{
 set.seed(45056)
