+++ DESCRIPTION
@@ -17,7 +17,7 @@ Imports: deldir, sp, rgeos, raster, methods
 License: MIT + file LICENSE
 Encoding: UTF-8
 LazyData: true
-RoxygenNote: 6.0.1.9000
+RoxygenNote: 6.1.1
 Suggests: knitr, rmarkdown, ggmap, dplyr, scales,
     testthat, vdiffr
 VignetteBuilder: knitr
