### 2018-11-26 garretrc/ggvoronoi#3 @garretrc
Added hyperlinks to all the documentation pages, including a link to the website in ncdc_locations!
