### 2018-11-26 2b2821c
added further information for oxford_bikes data
