### 2018-11-26 garretrc/ggvoronoi#5 @garretrc
Consulted with the people who gathered the data points and updated the documentation
