### 2018-11-26 #1096 @garretrc
@corybrunson I've addressed each issue, one is still open and I've left a comment as justification. If you accept that reasoning you can close the issue, otherwise leave a comment and I can make further changes. Feel free to re-open any of the other issues if you feel my solutions aren't acceptable! Thank you.
