### 2018-11-26 garretrc/ggvoronoi#4 @corybrunson
Oh! I see—`geom_voronoi` produces a layer using `GeomPolygon`! I had noticed that earlier but somehow forgot it when raising the issue.

I have a personal qualm with having a layer without an associated ggproto object, but perhaps that's not uncommon and i just haven't noticed it. I also got into the habit of calling `geom_*` by default, though recently i've tried to get into the habit of calling whichever layer function better communicates (to the code reader) the purpose of the resulting graphic.

But there's certainly room for different conventions here, and yours might ultimately be more useful than mine. And i see that it's not an oversight, so it no longer holds up my review. Thanks for the explanation!
