### 2018-11-26 #1096 @corybrunson
@garretrc thanks for addressing all of the issues! You've resolved or explained them to my satisfaction.

@yochannah i have completed my review and, conditional on my co-reviewer's feedback, recommend acceptance of the current GitHub version (https://github.com/garretrc/ggvoronoi/commit/2b2821cb71c3744e35ecfe6e69a9c20adffbf339).
