### 2018-11-26 garretrc/ggvoronoi#4 @garretrc
Thanks! I'll close the issue. This was a source of weeks of debate for us when we were first creating the package, resulting in a number of conversations with @kmaurer, my data viz prof and a contributor to the package, about whether or not to include geom_voronoi. Ultimately, we decided to keep it in order to have a more familiar syntax (like geom_point and geom_line etc.).

Early on in development when I was reading Hadley's guide to extending ggplot2 (http://rstudio-pubs-static.s3.amazonaws.com/108934_8537676801dd4548a96f6451bae01e94.html) he said something regarding having geom_* and stat_* instead of layer_* was a "mistake", and since reading that I've been more willing to choose user convenience over what is most likely "correct" (which would be having only stat_voronoi and no geom_*). Overall I think we made a good choice, but I do tend to use stat_voronoi over geom_voronoi personally!
