### 2018-12-03 40dfb68
Updated installation portion of README
