### 2018-12-03 c9ba48f
added hyperlinks to R packages in README.md
