### 2018-12-03 #1096 @trinker
@garretrc Thanks for addressing the couple of GitHub issues. These have been addressed/resolved.

@yochannah I have also completed my review and also recommend acceptance of the current GitHub version (garretrc/ggvoronoi@c9ba48f).
