### 2018-12-03 2de1266
Added wiki link for Voronoi diagram description
