### 2018-12-03 314a890
hyperlinks in Contributing.md
