+++ CONTRIBUTING.md
@@ -1,19 +1,19 @@
 # Contributing to ggvoronoi
 
-Have a good idea for ggvoronoi? Let's work together! First, please check out the [code of conduct.](https://github.com/garretrc/ggvoronoi/blob/master/CODE_OF_CONDUCT.md)
+Have a good idea for 'ggvoronoi'? Let's work together! First, please check out the [code of conduct.](CODE_OF_CONDUCT.md)
 
 ## Procdeures
 
 * Before submitting a pull request, please contact Robert Garrett at garretrc@miamioh.edu to discuss the error or idea. 
 * If you have found a bug, please include information about your R version and system OS. 
 * For each pull request you want to make, please create a new branch. 
-* Please include tests for each edit or new contribution made. Testing in 'ggvoronoi' is done using 'testthat' and 'vdiffr'.
+* Please include tests for each edit or new contribution made. Testing in ['ggvoronoi'](https://cran.r-project.org/package=ggvoronoi) is done using ['testthat'](https://cran.r-project.org/package=testthat) and ['vdiffr'](https://cran.r-project.org/package=vdiffr).
 
 ## Contributions
 
 Robert Garrett: Package maintainer, co-creator, and co-author
 * Co-author of base functionality in `voronoi_polygon`
-* Co-author of 'ggplot2' integration
+* Co-author of ['ggplot2'](https://cran.r-project.org/package=ggplot2) integration
 * Co-author of vignette
 * Co-author of documentation
 * Author of `fortify_voronoi`
@@ -21,7 +21,7 @@ Robert Garrett: Package maintainer, co-creator, and co-author
 
 Austin Nar: Co-creator and co-author
 * Co-author of base functionality in `voronoi_polygon`
-* Co-author of 'ggplot2' integration
+* Co-author of ['ggplot2'](https://cran.r-project.org/package=ggplot2) integration
 
 Thomas Fisher: Co-author
 * Co-author of vignette
@@ -29,4 +29,4 @@ Thomas Fisher: Co-author
 
 Karsten Maurer: Contributor
 * Contributions to package ideation
-* Contributions to 'ggplot2' integration
+* Contributions to ['ggplot2'](https://cran.r-project.org/package=ggplot2) integration
